+++ src/Symfony/Component/ClassLoader/CHANGELOG.md
@@ -4,8 +4,7 @@ CHANGELOG
 3.3.0
 -----
 
- * The ApcClassLoader, WinCacheClassLoader and XcacheClassLoader classes have been deprecated
-   in favor of the `--apcu-autoloader` option introduced in composer 1.3
+ * deprecated the component: use Composer instead
 
 3.0.0
 -----